lib/ukpal: Remove `lcpu.h` header and adapt users accordingly
The `lcpu.h` is an unpleasant reference to a library the PAL is not
supposed to be aware of and whose existence is no longer necessary:
it only contains the IPI sending prototype. Therefore, remove it
and move the IPI sending prototype to the exception namespace, as it
essentially implies synchronously sending an exception to another CPU
anyway.

Finally, adapt the rest of the codebase accordingly:
- lib/uklcpu no longer needs to import the pal/lcpu.h headers and
uk_pal_send_ipi is renamed to uk_pal_except_send_ipi to reflect the
namespace update
- the platform PAL driver's exception code cand take ownership of
IPI sending as well as remove related lcpu.h registrations
