+++ plat/native/exportsyms.uk
@@ -34,7 +34,7 @@ uk_plat_native_attr_from_pte
 uk_plat_native_x86_64_pg_maxphysaddr
 uk_plat_native_rdfsbasefn
 uk_plat_native_wrfsbasefn
-uk_plat_native_send_ipi
+uk_plat_native_except_send_ipi
 uk_plat_native_sentry
 uk_plat_native_sstackp
 uk_plat_native_sarg
